Enhance CSV output handling in SimulationResult by ensuring RFC 4180 compliance and using invariant culture for numeric values

+++ SimNextgenApp/Modeling/Generator/Generator.cs
@@ -101,7 +101,7 @@ public class Generator<TLoad> : AbstractSimulationModel, IGenerator<TLoad>, IOpe
     /// </remarks>
     public override void WarmedUp(long simulationTime)
     {
-        StartTime = (long)simulationTime;
+        StartTime = simulationTime;
         LoadsGeneratedCount = 0;
     }
